+++ src/main/java/org/c4marathon/assignment/domain/account/entity/ScheduleCreateEvent.java
@@ -1,6 +1,6 @@
 package org.c4marathon.assignment.domain.account.entity;
 
-import org.c4marathon.assignment.domain.user.entity.User;
+import org.c4marathon.assignment.domain.user.entity.user.User;
 
 public record ScheduleCreateEvent(User user) {
