+++ src/main/java/org/c4marathon/assignment/domain/settlement/entity/settlement/SettlementUser.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.domain.settlement.entity.settlement;
 
 import static jakarta.persistence.EnumType.*;
 
-import org.c4marathon.assignment.domain.user.entity.User;
+import org.c4marathon.assignment.domain.user.entity.user.User;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
