+++ src/main/java/org/c4marathon/assignment/domain/user/dto/UserMapper.java
@@ -1,7 +1,7 @@
 package org.c4marathon.assignment.domain.user.dto;
 
 import org.c4marathon.assignment.domain.user.dto.request.UserDto;
-import org.c4marathon.assignment.domain.user.entity.User;
+import org.c4marathon.assignment.domain.user.entity.user.User;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import lombok.AccessLevel;
