+++ src/main/java/org/c4marathon/assignment/domain/user/repository/UserRepository.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.domain.user.repository;
 
 import java.util.Optional;
 
-import org.c4marathon.assignment.domain.user.entity.User;
+import org.c4marathon.assignment.domain.user.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
