+++ src/main/java/org/c4marathon/assignment/domain/user/service/UserService.java
@@ -1,6 +1,6 @@
 package org.c4marathon.assignment.domain.user.service;
 
-import static org.c4marathon.assignment.domain.user.entity.UserErrCode.*;
+import static org.c4marathon.assignment.domain.user.entity.responsemsg.UserErrCode.*;
 
 import org.c4marathon.assignment.domain.account.entity.ScheduleCreateEvent;
 import org.c4marathon.assignment.domain.user.dto.response.JoinResponseDto;
@@ -8,9 +8,9 @@ import org.c4marathon.assignment.domain.user.dto.request.LoginRequestDto;
 import org.c4marathon.assignment.domain.user.dto.response.LoginResponseDto;
 import org.c4marathon.assignment.domain.user.dto.request.UserDto;
 import org.c4marathon.assignment.domain.user.dto.UserMapper;
-import org.c4marathon.assignment.domain.user.entity.JoinResponseMsg;
-import org.c4marathon.assignment.domain.user.entity.LoginResponseMsg;
-import org.c4marathon.assignment.domain.user.entity.User;
+import org.c4marathon.assignment.domain.user.entity.responsemsg.JoinResponseMsg;
+import org.c4marathon.assignment.domain.user.entity.responsemsg.LoginResponseMsg;
+import org.c4marathon.assignment.domain.user.entity.user.User;
 import org.c4marathon.assignment.domain.user.exception.UserException;
 import org.c4marathon.assignment.domain.user.repository.UserRepository;
 import org.springframework.context.ApplicationEventPublisher;
