+++ src/main/java/org/c4marathon/assignment/global/exception/ErrDtoMapper.java
@@ -5,7 +5,7 @@ import org.c4marathon.assignment.domain.account.entity.responseMsg.AccountErrCod
 import org.c4marathon.assignment.domain.settlement.dto.response.SettlementErrDto;
 import org.c4marathon.assignment.domain.settlement.entity.responsemsg.SettlementErrCode;
 import org.c4marathon.assignment.domain.user.dto.response.UserErrDto;
-import org.c4marathon.assignment.domain.user.entity.UserErrCode;
+import org.c4marathon.assignment.domain.user.entity.responsemsg.UserErrCode;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
